Move enum middleware into compat module
