+++ graphene/types/tests/test_enum.py
@@ -1,9 +1,7 @@
 from textwrap import dedent
-from graphql import OperationType
 
-from graphene.utils.str_converters import to_snake_case
+from graphene.compat.middleware import enum_value_convertor_middleware
 from ..argument import Argument
-from ..definitions import GrapheneEnumType
 from ..enum import Enum, PyEnum
 from ..field import Field
 from ..inputfield import InputField
@@ -398,12 +396,6 @@ def test_enum_mutation():
     assert results.data["setFavColor"]["favColor"] == Color.RED.name
 
 
-def get_underlying_type(_type):
-    while hasattr(_type, "of_type"):
-        _type = _type.of_type
-    return _type
-
-
 def test_enum_mutation_compat():
     from enum import Enum as PyEnum
 
@@ -436,23 +428,6 @@ def test_enum_mutation_compat():
     class MyMutations(ObjectType):
         set_fav_color = SetFavColor.Field()
 
-    def enum_compat_middleware(next, root, info, **args):
-        operation = info.operation.operation
-        if operation == OperationType.MUTATION:
-            input_arguments = info.parent_type.fields[info.field_name].args
-            for arg_name, arg in input_arguments.items():
-                _type = get_underlying_type(arg.type)
-                if isinstance(_type, GrapheneEnumType):
-                    # Convert inputs to value
-                    arg_name = to_snake_case(arg_name)
-                    input_value = args.get(arg_name, None)
-                    if input_value and isinstance(
-                        input_value, _type.graphene_type._meta.enum
-                    ):
-                        args[arg_name] = args[arg_name].value
-
-        return next(root, info, **args)
-
     schema = Schema(query=Query, mutation=MyMutations)
 
     results = schema.execute(
@@ -461,7 +436,7 @@ def test_enum_mutation_compat():
                 favColor
             }
         }""",
-        middleware=[enum_compat_middleware],
+        middleware=[enum_value_convertor_middleware],
     )
     assert not results.errors
